# api/openapi.yaml
openapi: 3.0.3
info:
  title: giro
  version: "1.0.0"
  description: |
    A double entry ledger.

    Value moves between accounts as **postings**. A posting names both sides and
    carries a positive amount, so an unbalanced entry cannot be expressed. A
    **transaction** is an ordered list of postings applied atomically.

    Money enters the ledger from the reserved `world` account and leaves to it.
    `world` is the only account permitted a negative balance, so for any asset
    every balance summed together is always exactly zero.

    ### Two clocks

    A transaction carries `timestamp`, when it happened economically, and
    `insertedAt`, when this database learned of it. They disagree whenever news
    arrives late, which for anything touching the outside world is most of the
    time.

    So there are two views of an account. What it holds **now**, and what it
    held **as of a date**, which is what `?at=` asks for. The second changes
    when a transaction is backdated; the first does not.

    ### Amounts

    Amounts are arbitrary precision integers in the asset's smallest unit, sent
    as JSON numbers. `USD/2` means two decimal places, so `10000` is $100.00.

    They can exceed 2^53. JavaScript's `JSON.parse` silently loses precision
    above that, so a browser client must use a big number aware parser rather
    than the built in one.
servers:
  - url: /
...
paths:
  /v1/ledgers/{ledger}:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
    get:
      tags: [ledgers]
      summary: Read a ledger
      operationId: getLedger
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [ledgers]
      summary: Create a ledger
      operationId: createLedger
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ledger" }
        "409": { $ref: "#/components/responses/Conflict" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/transactions:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
    get:
      tags: [transactions]
      summary: List transactions
      operationId: listTransactions
      parameters:
        - name: account
          in: query
          description: Only transactions where this address is a source or a destination.
          schema: { type: string }
        - name: accountPrefix
          in: query
          description: Only transactions touching an account under this prefix, for example `users:`.
          schema: { type: string }
        - name: reference
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransactionPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [transactions]
      summary: Commit a transaction
      description: |
        Applies every posting atomically, or none of them.

        Rejected if any account other than `world` would end below zero. The
        check is on the final state, so an account may pass through zero within
        a transaction.
      operationId: createTransaction
      parameters:
        - name: dryRun
          in: query
          description: |
            Run the whole commit path and then roll back, returning what would
            have happened.

            This is the real path rather than a simulation, so it cannot drift
            from a real commit: the locks are taken, the balances are checked
            against live data, and a transaction that would be rejected is
            rejected here too, with the same status.

            Nothing is consumed. No id is allocated and no idempotency key is
            claimed, so the id on the response is what it would have been rather
            than a reservation. Responds 200 rather than 201, because nothing
            was created.
          schema: { type: boolean, default: false }
        - name: Idempotency-Key
          in: header
          description: |
            Replaying this key returns the original transaction rather than
            creating a second one.

            The key is stored with a hash of the request body. Reusing a key
            with different inputs is an error rather than a silent success for
            a payment that never happened.
          schema: { type: string, maxLength: 256 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTransactionRequest" }
      responses:
        "200":
          description: What a dry run would have done. Nothing was committed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "201":
          description: Committed, or replayed from an earlier identical request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: The postings are valid but cannot be applied.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/TransactionID"
    get:
      tags: [transactions]
      summary: Read a transaction
      operationId: getTransaction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/transactions/{id}/revert:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/TransactionID"
    post:
      tags: [transactions]
      summary: Revert a transaction
      description: |
        Commits a new transaction holding the original's postings with both
        sides swapped. The original is never edited or deleted: it keeps its
        postings and gains a `revertedAt` mark saying a correction exists.

        The reversal is an ordinary transaction with its own id, and it can
        fail. If the money has since been spent it is not there to give back,
        and the reversal is rejected for insufficient funds like any other
        transaction would be.

        Balances return to where they were. Volumes do not: both counters go up,
        so the rows record that money moved and then came back, rather than
        looking like it never moved.
      operationId: revertTransaction
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RevertRequest" }
      responses:
        "201":
          description: Reverted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reversal" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already reverted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: The reversal cannot be applied, usually because the money has been spent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/transactions/{id}/metadata:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/TransactionID"
    post:
      tags: [transactions]
      summary: Merge metadata into a transaction
      description: |
        Merges the given keys into whatever is already there. Existing keys not
        mentioned are left alone, so two callers writing different keys do not
        overwrite each other.

        Writing metadata that is already present changes nothing and appends no
        log entry.
      operationId: setTransactionMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Metadata" }
      responses:
        "200":
          description: The transaction, with its metadata as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/transactions/{id}/metadata/{key}:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/TransactionID"
      - $ref: "#/components/parameters/MetadataKey"
    delete:
      tags: [transactions]
      summary: Remove one metadata key from a transaction
      description: Removing a key that is not there changes nothing and is not an error.
      operationId: deleteTransactionMetadata
      responses:
        "200":
          description: The transaction, with its metadata as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/accounts/{address}/moves:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/Address"
    get:
      tags: [accounts]
      summary: An account's statement
      description: |
        Every movement this account was part of, oldest first, in **effective
        date** order.

        That is what a statement means: the sequence in which things happened,
        not the sequence in which this database heard about them. The two differ
        whenever a transaction has been backdated.

        Each row carries the account's balance immediately after that movement,
        in both orderings. `volumes` is frozen at commit; `effectiveVolumes` is
        what the running balance actually was on that date.
      operationId: listMoves
      parameters:
        - name: asset
          in: query
          schema: { type: string }
        - name: from
          in: query
          description: Inclusive lower bound on the effective date.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: Inclusive upper bound on the effective date.
          schema: { type: string, format: date-time }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MovePage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/accounts/{address}/metadata:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/Address"
    post:
      tags: [accounts]
      summary: Merge metadata into an account
      description: |
        Creates the account if it does not exist. Accounts are never registered,
        so tagging one before any money has moved through it is normal.

        Merges rather than replaces, and writing metadata that is already
        present appends no log entry.
      operationId: setAccountMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Metadata" }
      responses:
        "200":
          description: The account, with its metadata as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/accounts/{address}/metadata/{key}:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/Address"
      - $ref: "#/components/parameters/MetadataKey"
    delete:
      tags: [accounts]
      summary: Remove one metadata key from an account
      operationId: deleteAccountMetadata
      responses:
        "200":
          description: The account, with its metadata as it now stands.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/accounts/{address}:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/Address"
    get:
      tags: [accounts]
      summary: Read an account
      description: |
        Accounts are never registered. One exists because a posting named it,
        so this returns 404 only for an address that has never been used.
      operationId: getAccount
      parameters:
        - name: expand
          in: query
          description: |
            Comma separated extras to include: `volumes`, `effectiveVolumes`.
            Omitted by default because each costs a second query.

            `volumes` is what the account holds now. `effectiveVolumes` is what
            it held as of `at`, which differs whenever a transaction has been
            backdated.
          schema: { type: string }
        - $ref: "#/components/parameters/At"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/accounts/{address}/balances:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
      - $ref: "#/components/parameters/Address"
    get:
      tags: [accounts]
      summary: Read one account's balances
      description: |
        An address that has never been touched returns an empty object rather
        than a 404: its balance is zero either way.
      operationId: getBalances
      parameters:
        - $ref: "#/components/parameters/At"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balances" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/balances:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
    get:
      tags: [accounts]
      summary: Aggregate balances across a subtree
      description: |
        Sums every account under a prefix. With no prefix this covers the whole
        ledger, including `world`, and must therefore be exactly zero for every
        asset. That is the conservation invariant, readable over HTTP.
      operationId: aggregateBalances
      parameters:
        - name: prefix
          in: query
          description: For example `users:`. Matches on segment boundaries, so `users:` never matches `fees:users:refunds`.
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balances" }
        default: { $ref: "#/components/responses/Error" }
  /v1/ledgers/{ledger}/logs:
    parameters:
      - $ref: "#/components/parameters/LedgerPath"
    get:
      tags: [logs]
      summary: List log entries
      description: |
        The log is the source of truth. Every mutation appends an entry whose
        hash covers the previous entry's hash, so editing history invalidates
        every hash after the edit.

        Ids are gapless, so a missing entry is detectable.
      operationId: listLogs
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LogPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        default: { $ref: "#/components/responses/Error" }
components:
  parameters:
...
    LedgerPath:
      name: ledger
      in: path
      required: true
      description: Ledger name. Each ledger is a separate set of books with its own ids and its own log chain.
      schema: { type: string }
    TransactionID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    At:
      name: at
      in: query
      description: |
        An effective date. Answers what was true on that date rather than what
        is true now, which differs whenever a transaction has been backdated.

        Defaults to now.
      schema: { type: string, format: date-time }
    MetadataKey:
      name: key
      in: path
      required: true
      description: A single metadata key.
      schema: { type: string, maxLength: 128 }
    Address:
      name: address
      in: path
      required: true
      description: Account address, colon separated segments, for example `users:alice:wallet`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque position from a previous response. When present every other
        query parameter is ignored, because the cursor carries the filter it
        was created with.
      schema: { type: string }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflicts with something that already exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Error:
      description: Unexpected error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Balances:
      type: object
      description: Asset to balance. A balance may be negative only for `world`.
      additionalProperties: { $ref: "#/components/schemas/Amount" }
      example: { "USD/2": 7500 }
    Metadata:
      type: object
      description: |
        Opaque to the ledger. Never interpreted, only stored.

        At most 32 keys, keys up to 128 bytes and values up to 1024 bytes. It is
        for identifiers and short labels; anything larger belongs in whatever
        system owns the document.
      maxProperties: 32
      additionalProperties: { type: string, maxLength: 1024 }
      example: { "orderId": "ord_8812", "provider": "stripe" }
    CreateTransactionRequest:
      type: object
      required: [postings]
      properties:
        postings:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Posting" }
        timestamp:
          type: string
          format: date-time
          description: |
            When the movement happened economically, which may be earlier than
            now. Defaults to now. Stored at microsecond precision.
        reference:
          type: string
          description: The caller's own identifier, unique per ledger.
        metadata: { $ref: "#/components/schemas/Metadata" }
    Transaction:
      type: object
      required: [id, postings, timestamp, insertedAt]
      properties:
        id: { type: integer, format: int64 }
        postings:
          type: array
          items: { $ref: "#/components/schemas/Posting" }
        timestamp:
          type: string
          format: date-time
          description: When it happened economically.
        insertedAt:
          type: string
          format: date-time
          description: When this database learned of it. May be later than `timestamp`.
        revertedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when a reversal exists. The reversal is a separate transaction with its own id.
        reference: { type: string }
        metadata: { $ref: "#/components/schemas/Metadata" }
        postCommitVolumes:
          type: object
          description: Volumes of every touched account immediately after this transaction. Frozen at commit and never rewritten.
          additionalProperties:
            type: object
            additionalProperties: { $ref: "#/components/schemas/Volumes" }
...
    RevertRequest:
      type: object
      properties:
        atEffectiveDate:
          type: boolean
          default: false
          description: |
            Date the reversal with the original's effective date rather than
            now. Off by default: a reversal happens when it happens, and
            backdating one rewrites what historical balances say about a period
            that has probably already been reported on.
        force:
          type: boolean
          default: false
          description: |
            Commit even if an account other than `world` ends below zero.

            A reversal can legitimately fail. This manufactures a negative
            balance instead, and exists for an operator who has decided that is
            the lesser problem.
    Reversal:
      type: object
      required: [original, reversal]
      description: The pair, so a caller does not have to fetch the other half.
      properties:
        original:
          allOf:
            - $ref: "#/components/schemas/Transaction"
          description: The transaction that was undone, now carrying `revertedAt`.
        reversal:
          allOf:
            - $ref: "#/components/schemas/Transaction"
          description: The transaction that undid it, tagged with `giro/reverts`.
    Ledger:
      type: object
      required: [name, addedAt]
      properties:
        name: { type: string }
        addedAt: { type: string, format: date-time }
...
    MovePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Move" }
        next: { type: string }
    TransactionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Transaction" }
        next:
          type: string
          description: Cursor for the following page. Absent when there is nothing after this one.
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable identifier. Branch on this rather than on the message.
          enum:
            - VALIDATION
            - NOT_FOUND
            - CONFLICT
            - INSUFFICIENT_FUNDS
            - IDEMPOTENCY_MISMATCH
            - INTERNAL
        message: { type: string }
        details:
          type: object
          additionalProperties: true
          description: Extra context. For INSUFFICIENT_FUNDS this carries the account, asset, available and requested.